fix: keep rsc context exports server-safe
