+++ packages/next/src/index.server.ts
@@ -36,7 +36,6 @@ import {
   useGT,
 } from './server-dir/buildtime/getTranslationFunction';
 import type { LocaleProperties } from '@generaltranslation/format/types';
-export { LocaleSelector, RegionSelector } from './index.client';
 
 export function useGTClass() {
   return getI18NConfig().getGTClass();
