+++ packages/react-core/src/components/branches/Branch.tsx
@@ -1,35 +1,14 @@
 import type { ReactNode } from 'react';
-
-type BranchProps = {
-  children?: ReactNode;
-  branch?: string | number | boolean;
-  _locale?: string;
-  _enableI18n?: boolean;
-  [key: string]: ReactNode;
-};
+import { computeBranch } from './computeBranch';
+import type { BranchProps } from './computeBranch';
 
 // ===== Component ===== //
 
 /**
  * External-store version of the `<Branch>` component.
  */
-function GtInternalBranch({
-  children,
-  branch,
-  _locale,
-  _enableI18n,
-  ...branches
-}: BranchProps): ReactNode {
-  void _locale;
-  void _enableI18n;
-
-  let branchKey = branch?.toString();
-  if (typeof branchKey === 'string' && branchKey.startsWith('data-')) {
-    branchKey = undefined;
-  }
-  return branchKey && typeof branches[branchKey] !== 'undefined'
-    ? branches[branchKey]
-    : children;
+function GtInternalBranch(props: BranchProps): ReactNode {
+  return computeBranch(props);
 }
 
 function Branch(props: BranchProps): React.JSX.Element {
