+++ packages/react-core/src/components/branches/Plural.tsx
@@ -1,34 +1,22 @@
-import getPluralBranch from '../../utils/plurals/getPluralBranch';
 import type { ReactNode } from 'react';
 import { useEnableI18n, useLocale } from '../../hooks/condition-store';
-import { getFormatLocales } from '../../hooks/utils';
-
-type PluralProps = {
-  children?: ReactNode;
-  n: number;
-  locales?: string[];
-  _locale?: string;
-  _enableI18n?: boolean;
-  [key: string]: ReactNode;
-};
+import { computePlural } from './computePlural';
+import type { PluralProps } from './computePlural';
 
 // ===== Component ===== //
 
 function GtInternalPlural({
   _enableI18n,
   _locale,
-  children,
-  n,
-  locales: localesProp = [],
-  ...branches
+  ...props
 }: PluralProps): ReactNode {
   const locale = _locale ?? useLocale();
   const enableI18n = _enableI18n ?? useEnableI18n();
-  const locales = getFormatLocales({ locale, enableI18n, localesProp });
-  if (typeof n !== 'number') {
-    return children;
-  }
-  return getPluralBranch(n, locales, branches) || children;
+  return computePlural({
+    ...props,
+    _locale: locale,
+    _enableI18n: enableI18n,
+  });
 }
 
 function Plural(props: PluralProps): React.JSX.Element {
