+++ packages/react-core/src/components/variables/Num.tsx
@@ -1,41 +1,6 @@
 import { useEnableI18n, useLocale } from '../../hooks/condition-store';
-import { getFormatLocales } from '../../hooks/utils';
-import { getI18nConfig } from 'gt-i18n/internal';
-
-type NumProps = {
-  children: number | string | null | undefined;
-  options?: Intl.NumberFormatOptions;
-  locales?: string[];
-  name?: string;
-  _locale?: string;
-  _enableI18n?: boolean;
-};
-
-type ResolvedNumProps = NumProps & {
-  _locale: string;
-  _enableI18n: boolean;
-};
-
-// ===== Shared Logic ===== //
-
-function computeNum({
-  _enableI18n,
-  _locale,
-  children,
-  options = {},
-  locales: localesProp = [],
-}: ResolvedNumProps): string | null {
-  const locales = getFormatLocales({
-    locale: _locale,
-    enableI18n: _enableI18n,
-    localesProp,
-  });
-  const gt = getI18nConfig().getGTClass();
-  if (children == null) return null;
-  const parsedNumber =
-    typeof children === 'string' ? parseFloat(children) : children;
-  return gt.formatNum(parsedNumber, { locales, ...options });
-}
+import { computeNum } from './computeNum';
+import type { NumProps } from './computeNum';
 
 // ===== Component ===== //
