+++ packages/react-core/src/context.ts
@@ -81,22 +81,3 @@ export type {
 } from './hooks/utils/missing-translation';
 export type { TranslateLookup } from './i18n-store/storeTypes';
 export type { DictionaryLookup } from './i18n-store/storeTypes';
-export { GtInternalBranch } from './components/branches/Branch';
-export { GtInternalPlural } from './components/branches/Plural';
-export { GtInternalDerive } from './components/derivation/Derive';
-export { GtInternalCurrency } from './components/variables/Currency';
-export { GtInternalDateTime } from './components/variables/DateTime';
-export { GtInternalNum } from './components/variables/Num';
-export { GtInternalRelativeTime } from './components/variables/RelativeTime';
-export { GtInternalVar } from './components/variables/Var';
-export { renderPreparedT } from './utils/rendering/renderPreparedT';
-export { prepareT, usePrepareT } from './utils/translation/prepareT';
-export type {
-  JsxTranslationOptions,
-  PreparedT,
-} from './utils/translation/prepareT';
-export type {
-  TaggedChildren,
-  RelativeTimeFormatOptions,
-  RenderVariable,
-} from './utils/types';
