+++ packages/react-core/src/functions/translation/t.ts
@@ -15,7 +15,7 @@ import {
 } from '../../condition-store/singleton-operations';
 import { StringContent, StringFormat } from 'generaltranslation/types';
 import { getReactI18nCache } from '../../i18n-cache/singleton-operations';
-import { getShouldTranslate } from '../../hooks/utils';
+import { getShouldTranslate } from '../../hooks/should-translate';
 import { createDiagnosticMessage } from 'generaltranslation/internal';
 
 /**
