+++ packages/react-core/src/utils/rendering/__tests__/renderVariable.test.tsx
@@ -99,6 +99,7 @@ describe('renderVariable locale handling', () => {
       children: createNumberVariable(5),
       defaultLocale: 'en',
       enableI18n: false,
+      renderVariable,
     });
 
     expect(React.isValidElement(result)).toBe(true);
@@ -117,6 +118,7 @@ describe('renderVariable locale handling', () => {
       target: { k: 'count', v: 'n' },
       locales: ['fr', 'en'],
       enableI18n: true,
+      renderVariable,
     });
 
     expect(React.isValidElement(result)).toBe(true);
