+++ packages/react/src/context.server.ts
@@ -1,26 +1,62 @@
 'server-only';
 
-export { initializeGTSPA } from './setup/initializeGTSPA';
-export { useLocaleSelector } from './components/useLocaleSelector';
-export { useSetLocale, useSetEnableI18n } from './hooks/conditions-store';
 export {
   defaultEnableI18nCookieName,
   defaultLocaleCookieName,
   defaultRegionCookieName,
 } from './cookie-names';
-export { RscBranch } from './components/branches/Branch';
-export { RscPlural } from './components/branches/Plural';
-export { RscDerive } from './components/derivation/Derive';
-export { RscT } from './components/translation/T';
-export { RscCurrency } from './components/variables/Currency';
-export { RscDateTime } from './components/variables/DateTime';
-export { RscNum } from './components/variables/Num';
-export { RscRelativeTime } from './components/variables/RelativeTime';
-export { RscVar } from './components/variables/Var';
 
-// ===== Components ===== //
-export { ServerGTProvider as GTProvider } from './provider/ServerGTProvider';
-export { LocaleSelector } from './components/LocaleSelector';
+/**
+ * The following are client-only APIs. They are surfaced from the `/context`
+ * entry for type parity, but importing their real implementations would pull
+ * `createContext` into the server (RSC) module graph, so they throw if called
+ * from the server condition.
+ */
+function throwServerOnlyBoundaryError(name: string): never {
+  throw new Error(
+    `gt-react: ${name} is not available from the react-server context entry.`
+  );
+}
+
+export function GTProvider(): never {
+  return throwServerOnlyBoundaryError('GTProvider');
+}
+
+export function LocaleSelector(): never {
+  return throwServerOnlyBoundaryError('LocaleSelector');
+}
+
+export function initializeGTSPA(): never {
+  return throwServerOnlyBoundaryError('initializeGTSPA');
+}
+
+export function internalInitializeGTSPA(): never {
+  return throwServerOnlyBoundaryError('internalInitializeGTSPA');
+}
+
+export function useLocaleSelector(): never {
+  return throwServerOnlyBoundaryError('useLocaleSelector');
+}
+
+export function useSetLocale(): never {
+  return throwServerOnlyBoundaryError('useSetLocale');
+}
+
+export function useSetEnableI18n(): never {
+  return throwServerOnlyBoundaryError('useSetEnableI18n');
+}
+
+export function useGT(): never {
+  return throwServerOnlyBoundaryError('useGT');
+}
+
+export function useMessages(): never {
+  return throwServerOnlyBoundaryError('useMessages');
+}
+
+export function useTranslations(): never {
+  return throwServerOnlyBoundaryError('useTranslations');
+}
 
 export {
   // ===== Components ===== //
@@ -28,6 +64,7 @@ export {
   Plural,
   Derive,
   GtInternalTranslateJsx,
+  RscT,
   T,
   Currency,
   DateTime,
@@ -42,9 +79,6 @@ export {
   useLocales,
   getFormatLocales,
   useFormatLocales,
-  useGT,
-  useMessages,
-  useTranslations,
   // ===== Functions ===== //
   msg,
   decodeMsg,
@@ -60,5 +94,4 @@ export {
   t,
   // ===== Setup ===== //
   internalInitializeGTSRA as initializeGT,
-  internalInitializeGTSPA,
-} from '@generaltranslation/react-core/context';
+} from '@generaltranslation/react-core/context-server';
